Island: Move UserCreds from environment -> services/authentication
