+++ monkey/monkey_island/cc/services/authentication/i_user_datastore.py
@@ -1,6 +1,6 @@
 import abc
 
-from monkey_island.cc.environment.user_creds import UserCreds
+from .user_creds import UserCreds
 
 
 class IUserDatastore(metaclass=abc.ABCMeta):
