+++ monkey/monkey_island/cc/services/authentication/user_creds.py
