+++ monkey/tests/unit_tests/monkey_island/cc/services/authentication/test_authentication_service.py
@@ -8,10 +8,10 @@ from common.utils.exceptions import (
     InvalidRegistrationCredentialsError,
     UnknownUserError,
 )
-from monkey_island.cc.environment.user_creds import UserCreds
 from monkey_island.cc.services import AuthenticationService
 from monkey_island.cc.services.authentication import authentication_service
 from monkey_island.cc.services.authentication.i_user_datastore import IUserDatastore
+from monkey_island.cc.services.authentication.user_creds import UserCreds
 
 USERNAME = "user1"
 PASSWORD = "test"
