+++ monkey/tests/unit_tests/monkey_island/cc/services/authentication/test_user_creds.py
@@ -1,4 +1,4 @@
-from monkey_island.cc.environment.user_creds import UserCreds
+from monkey_island.cc.services.authentication.user_creds import UserCreds
 
 TEST_USER = "Test"
 TEST_HASH = "abc1231234"
